refactor(database): split drill_critical_tables.go out of drill.go (300-line cap)
The critical-tables resolution logic added in the prior commit pushed
drill.go to 322 lines, tripping internal/repoqa's TestFileSizeBudgetNotExceeded
(engineering-standard file-size cap, budget 0 over 300). Moves
DefaultCriticalTables + ResolveCriticalTables into their own file, following
the existing drill_*.go concern-split convention. No behavior change.
